requirements.txt
openai-whisper
pyannote.audio
pyaudio
Jinja2
ollama
